requirements-django.txt
https://www.djangoproject.com/download/1.7b3/tarball/
psycopg2
djorm-ext-pgarray
jsonfield
django-uuidfield
